Shop
Shop for stats

+++ Assets/Script/MenuScript/ShopScript/ShopItem.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 
     public string name;
 
-    public float price;
+    public int price;
 }
